pyproject.toml
[project]
name = "sangsassistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.2.1",
    "openai>=1.84.0",
    "pytest>=8.4.0",
    "python-dotenv>=1.1.0",
]
